requirements.txt
numpy
pandas
pymysql
python-dotenv
mysql-connector-python
scikit-learn


# -e .